+++ Parking_Be/app/Http/Controllers/CanHoController.php
@@ -10,6 +10,14 @@ class CanHoController extends Controller
 {
     public function getData()
     {
+        $id_chuc_nang = 3;
+        $check = $this->checkQuyen($id_chuc_nang);
+        if ($check == false) {
+            return response()->json([
+                'status'  =>  false,
+                'message' =>  'Bạn không có quyền chức năng này'
+            ]);
+        }
         $canho = CanHo::all();
         return response()->json([
             'status' => true,
@@ -17,6 +25,7 @@ class CanHoController extends Controller
             'data' => $canho
         ]);
     }
+    
     public function getDataClient()
     {
         $canho = CanHo::all();
@@ -29,11 +38,19 @@ class CanHoController extends Controller
 
     public function themCanHo(themCanHoRequest $request)
     {
+        $id_chuc_nang = 3;
+        $check = $this->checkQuyen($id_chuc_nang);
+        if ($check == false) {
+            return response()->json([
+                'status'  =>  false,
+                'message' =>  'Bạn không có quyền chức năng này'
+            ]);
+        }
         CanHo::create([
-           'tang'       => $request->tang,
-           'so_can_ho'  => $request->so_can_ho,
-           'chu_ho'     => $request->chu_ho,
-           'ten_toa_nha'     => $request->ten_toa_nha,
+            'tang'       => $request->tang,
+            'so_can_ho'  => $request->so_can_ho,
+            'chu_ho'     => $request->chu_ho,
+            'ten_toa_nha'     => $request->ten_toa_nha,
         ]);
         return response()->json([
             'status'   => true,
@@ -42,13 +59,21 @@ class CanHoController extends Controller
     }
     public function capnhatCanHo(Request $request)
     {
+        $id_chuc_nang = 3;
+        $check = $this->checkQuyen($id_chuc_nang);
+        if ($check == false) {
+            return response()->json([
+                'status'  =>  false,
+                'message' =>  'Bạn không có quyền chức năng này'
+            ]);
+        }
         $canho = CanHo::find($request->id);
         if ($canho) {
             $canho->update([
                 'tang'       => $request->tang,
                 'so_can_ho'  => $request->so_can_ho,
                 'chu_ho'     => $request->chu_ho,
-                 'ten_toa_nha'     => $request->ten_toa_nha,
+                'ten_toa_nha'     => $request->ten_toa_nha,
 
             ]);
             return response()->json([
@@ -64,6 +89,14 @@ class CanHoController extends Controller
     }
     public function xoaCanHo(Request $request)
     {
+        $id_chuc_nang = 3;
+        $check = $this->checkQuyen($id_chuc_nang);
+        if ($check == false) {
+            return response()->json([
+                'status'  =>  false,
+                'message' =>  'Bạn không có quyền chức năng này'
+            ]);
+        }
         $canho = CanHo::find($request->id);
         if ($canho) {
             $canho->delete();
