+++ Parking_Be/app/Http/Controllers/ViTriDatController.php
@@ -11,6 +11,14 @@ class ViTriDatController extends Controller
 {
     public function getData()
     {
+        $id_chuc_nang = 8;
+        $check = $this->checkQuyen($id_chuc_nang);
+        if ($check == false) {
+            return response()->json([
+                'status'  =>  false,
+                'message' =>  'Bạn không có quyền chức năng này'
+            ]);
+        }
         $vitri = ViTriDat::where('trang_thai', 1)->get();
         return response()->json([
             'data' => $vitri,
@@ -18,6 +26,14 @@ class ViTriDatController extends Controller
     }
     public function themViTriDat(themViTriDatRequest $request)
     {
+        $id_chuc_nang = 8;
+        $check = $this->checkQuyen($id_chuc_nang);
+        if ($check == false) {
+            return response()->json([
+                'status'  =>  false,
+                'message' =>  'Bạn không có quyền chức năng này'
+            ]);
+        }
         $vitri = ViTriDat::create([
             'vi_tri_dat' => $request->vi_tri_dat,
             'trang_thai' => 1,
@@ -30,6 +46,14 @@ class ViTriDatController extends Controller
     }
     public function capnhatViTriDat(capnhatViTriDat $request)
     {
+        $id_chuc_nang = 8;
+        $check = $this->checkQuyen($id_chuc_nang);
+        if ($check == false) {
+            return response()->json([
+                'status'  =>  false,
+                'message' =>  'Bạn không có quyền chức năng này'
+            ]);
+        }
         $vitri = ViTriDat::find($request->id);
         $vitri->update([
             'vi_tri_dat' => $request->vi_tri_dat,
@@ -42,6 +66,14 @@ class ViTriDatController extends Controller
     }
     public function xoaViTriDat($id)
     {
+        $id_chuc_nang = 8;
+        $check = $this->checkQuyen($id_chuc_nang);
+        if ($check == false) {
+            return response()->json([
+                'status'  =>  false,
+                'message' =>  'Bạn không có quyền chức năng này'
+            ]);
+        }
         $data = ViTriDat::find($id)->delete();
 
         return response()->json([
